requirements.txt
asttokens==2.4.1
backcall==0.2.0
bidict==0.23.1
blinker==1.7.0
Bootstrap-Flask==2.3.3
cffi==1.16.0
click==8.1.7
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
flask==2.3.3
Flask-JWT==0.3.2
Flask-SocketIO==5.3.6
h11==0.14.0
importlib-metadata==7.0.2
ipykernel==6.29.3
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter-client==8.6.0
jupyter-core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pyzmq==25.1.2
simple-websocket==1.0.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing-extensions==4.10.0
wcwidth==0.2.13
werkzeug==3.0.1
wsproto==1.2.0
wtforms==3.1.2
zipp==3.17.0
